face_emotion_recog_realtime/requirements.txt
flask
gunicorn
tensorflow-cpu
opencv-python-headless
numpy
